Decompose chessboard.rs
